+++ great_international/panels/great_international.py
@@ -1,7 +1,8 @@
 from django.forms import Textarea, CheckboxSelectMultiple
 
 from wagtail.admin.edit_handlers import (
-    HelpPanel, FieldPanel, FieldRowPanel, MultiFieldPanel, PageChooserPanel
+    InlinePanel, HelpPanel, FieldPanel, FieldRowPanel, MultiFieldPanel,
+    PageChooserPanel
 )
 from wagtail.images.edit_handlers import ImageChooserPanel
 
@@ -391,7 +392,7 @@ class InternationalHomePageOldPanels:
     )
 
 
-class InternationalSectorPagePanels:
+class BaseInternationalSectorPagePanels:
 
     content_panels = [
         MultiFieldPanel(
@@ -1023,3 +1024,113 @@ class InternationalEUExitFormSuccessPagePanels:
         FieldPanel('title_en_gb'),
         FieldPanel('slug'),
     ]
+
+
+class AboutDitLandingPagePanels:
+
+    content_panels = [
+        FieldPanel('breadcrumbs_label'),
+        MultiFieldPanel(
+            heading="Hero",
+            children=[
+                FieldPanel('hero_title'),
+                ImageChooserPanel('hero_image'),
+            ],
+        ),
+    ]
+
+    settings_panels = [
+        FieldPanel('title_en_gb'),
+        FieldPanel('slug'),
+        FieldPanel('uses_tree_based_routing'),
+    ]
+
+    edit_handler = make_translated_interface(
+        content_panels=content_panels,
+        settings_panels=settings_panels
+    )
+
+
+class AboutDitServiceFieldPanels:
+
+    panels = [
+        MultiFieldPanel([
+            ImageChooserPanel('icon'),
+            FieldPanel('title'),
+            FieldPanel('summary'),
+            FieldPanel('link_text'),
+            FieldPanel('link_url'),
+        ]),
+    ]
+
+
+class AboutDitServicesPagePanels:
+
+    content_panels = [
+        FieldPanel('breadcrumbs_label'),
+        MultiFieldPanel(
+            heading="Hero",
+            children=[
+                FieldPanel('hero_title'),
+                ImageChooserPanel('hero_image'),
+            ],
+        ),
+        MultiFieldPanel(
+            heading="Teaser",
+            children=[
+                HelpPanel('At least one field required for section to show'),
+                FieldPanel('teaser'),
+                ImageChooserPanel('teaser_image'),
+            ],
+        ),
+        MultiFieldPanel(
+            heading="Services section",
+            classname='collapsible',
+            children=[
+                InlinePanel(
+                    'about_dit_services_fields',
+                    label="About DIT services"
+                )
+            ]
+        ),
+        MultiFieldPanel(
+            heading="Case study",
+            classname='collapsible',
+            children=[
+                HelpPanel('Required fields for section to show: '
+                          'Case Study Image, Case Study Title'),
+                ImageChooserPanel('case_study_image'),
+                FieldPanel('case_study_title'),
+                FieldPanel('case_study_text'),
+                HelpPanel('CTAs require both text and a link to show '
+                          'on page. '),
+                FieldPanel('case_study_cta_text'),
+                FieldPanel('case_study_cta_link'),
+            ],
+        ),
+        MultiFieldPanel(
+            heading="Contact Section",
+            classname='collapsible',
+            children=[
+                HelpPanel('Required fields for section to show: '
+                          'Title, Summary'),
+                FieldPanel('contact_us_section_title'),
+                FieldPanel('contact_us_section_summary'),
+                HelpPanel('CTAs require both text and a link to show '
+                          'on page. '),
+                FieldPanel('contact_us_section_cta_text'),
+                FieldPanel('contact_us_section_cta_link'),
+            ],
+        ),
+    ]
+
+    settings_panels = [
+        FieldPanel('title_en_gb'),
+        FieldPanel('slug'),
+        FieldPanel('uses_tree_based_routing'),
+    ]
+
+    edit_handler = make_translated_interface(
+        content_panels=content_panels,
+        settings_panels=settings_panels
+    )