+++ CHANGELOG.md
@@ -10,6 +10,10 @@
 - CI-267 - Added cta text and link for sector page for related opportunities section
 - XOT-932 - Add Country Guide pages to search
 - CMS-1727 - Move region pages from Invest to International
+- CI-324 - Made `BaseInternationalSectorPage` and sector and new sub-sector page inherit from this and opportunity listing page filters by sub sector
+- CI-322 - Added "About DIT" placeholder landing page
+- CI-322 - Added "About DIT" services pages
+- CMS-1648 - Added tree-based routing helpers
 
 ### Fixed Bugs
 - NOTICKET - Capital invest landing page image serializers use aspect ratio for card images